+++ tests/Common/UserInfo.php
@@ -1,8 +1,10 @@
 <?php
 namespace Tests\Common;
 
+use Praline\Session\SessionDataInterface;
+
 // 如果測試過程中需要一些代表資料的小類別，可以用這個
-class UserInfo
+class UserInfo implements SessionDataInterface
 {
     /** @var  int */
     public $id;
@@ -14,5 +16,15 @@ class UserInfo
     {
         $this->id = $id;
         $this->name = $name;
+    }
+
+    public function getUniqueKey(): string
+    {
+        return strval($this->id);
+    }
+
+    public function getDescription(): string
+    {
+        return 'userId(' . strval($this->id) . ')';
     }
 }